// ANGULAR/ANGULAR/src/app/Shared/Model/commonddl.model.ts
export class DdlItemModel {
  Value: any;
  Text: string;
}